Rename 'feature' folder to 'features'
