+++ src/shared/api/api-context.tsx
@@ -4,7 +4,7 @@ import createClient from "openapi-fetch";
 import { createContext } from "react";
 import { paths } from "./generated/schema";
 // Note: Consider restructuring to avoid importing from a higher level
-import { ApiContextProps } from "@/feature/api-context-provider";
+import { ApiContextProps } from "@/features/api-context-provider";
 
 export const ApiContext = createContext<ApiContextProps>({} as ApiContextProps);
